requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pynput==1.7.7
pystray==0.19.5
pillow==10.4.0
sqlmodel==0.0.24
requests==2.32.3
pydantic==2.11.7
python-dotenv==1.1.1
